feat(bom-estimator): add shared help popup and drop legacy monolith
