+++ enrichment/__init__.py
@@ -1,0 +1,13 @@
+"""Generic part metadata enrichment providers."""
+
+from .providers import (
+    AssemblyMetadataProvider,
+    LCSCAssemblyMetadataProvider,
+    fetch_assembly_processes,
+)
+
+__all__ = [
+    "AssemblyMetadataProvider",
+    "LCSCAssemblyMetadataProvider",
+    "fetch_assembly_processes",
+]
